requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlmodel>=0.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.4.2
alembic>=1.12.0
python-dotenv>=1.0.0
httpx>=0.24.1
pytest>=7.4.2
pytest-asyncio>=0.21.1
# 数据可视化依赖 (用于API返回图表数据)
pandas>=2.0.0
matplotlib>=3.7.1
